Run usrsctp in single threaded mode.
With an option in PeerConnectionFactory to specify the thread. The plan
is that initially, passing nullptr will cause a new thread to be
created, and once users that need a standalone thread are updated, the
default will change to using the network thread.

When using a thread other than the network thread, any usrsctp operation
(including input and output) becomes a synchronous Thread::Invoke.
This is not ideal for performance, but should be a nonissue if that
mode of operation only ends up used for test frameworks, as we expect.

Change-Id: Ibdfb47c7dd137251214f324edf2dadb51b353c64