Fixing a buffer copy issue in DesktopFrame
This CL fixes a buffer copying issue introduced in this CL:
https://webrtc-review.googlesource.com/c/src/+/196485

In the BasicDesktopFrame::CopyOf function, the src and dst params
were swapped.  For me this manifested as a missing cursor when using
Chrome Remote Desktop.  I don't know of any other bugs this caused
but I have to assume it affects all callers of the function given
that the copy will never occur.

Bug: chromium:1197210
Change-Id: I166fcd3800f4e4bd8eb2d2b965dd194c4206215f