Use unordered map in RoundRobinPacketQueue
In highly loaded Media Servers, RoundRobinPacketQueue's usage of
std::map attributes to around 0.3% CPU. In profiling, it has been found
that it's the `streams_` map that is the culprit and std::map::find
being the function that is used most.

By using an unordered map, lookups should be faster. This will be
evaluated and if this commit doesn't show results, it will be reverted.

Bug: webrtc:12689
Change-Id: Ia368f1184130298cfbb9064528828435aa7a5c66