Use unordered_set in RTCPReceiver
In highly loaded media servers, RTCPReceiver's use of std::set
attributes to ~0.5% CPU. It's mostly ::find and the [] operator, and
they are all keyed by SSRC, which is an unordered data type. This makes
these maps suitable as unordered maps, as they have constant time
complexity for lookups.

After this commit has landed, it will be evaluated and if it doesn't
live up to expected savings, it will be reverted.

Question to the receiver: The copying of registered_ssrcs in
RTCPReceiver::TriggerCallbacksFromRtcpPacket seems costly, as later a
single SSRC is looked up. In large meetings, this table should be fairly
large. Could this be changed so that no copy is made?

Bug: webrtc:12689
Change-Id: I734578c22eeca2d9ba89fef77ecc689b72624567