Rename release_aar.py and modify it for validation of aar file.
As announced in the google groups [1], the pre-built Android aar is no
longer distributed and last update was August 2020. [2]
So we can remove the code that uploads aar to bintray in release_aar.py.

Still, the ability to create an Android aar and use it in a gradle
project (examples/aarproject) is useful. It can also be used to validate
aar by running PeerConnectionClientTest from examples/androidtests.

So I renamed release_aar.py to test_aar.py and make it working without
releasing the aar to an external hosting server.
This makes it easy to verify further changes to the aar.

[1] https://groups.google.com/g/discuss-webrtc/c/Ozvbd0p7Q1Y/m/TtQyRI1KAgAJ
[2] https://mvnrepository.com/artifact/org.webrtc/google-webrtc?repo=bt-google-webrtc

Bug: webrtc:11962
Change-Id: Ibe066a3a770569924e3b57805986808e1dd19df6