Remove UpdateReceiveHistograms and move the code to the dtor.
UpdateReceiveHistograms was only called from the dtor and moving the
code there allows us to use RTC_GUARDED_BY for member variables that
soon will be only accessed from the network thread (currently worker).

The dtor will continue to run on the worker though and while it has
a known execution context, UpdateReceiveHistograms does not and would
conflict with the network thread unless we move the code.

Bug: webrtc:11993
Change-Id: I363b4b342594f867f804972db99eede048ee3735