dcsctp: Add network/throughput tests
This initial version contains two tests, one when a lossy link is used
and one when there is no packet loss. It uses simulated time, but is
still only possible to run in optimized mode (without DCHECKS) as those
consistency checks inspect the outstanding data for inconsistencies on
each socket operation - which makes the test take too long time.

Bug: webrtc:12943
Change-Id: I9323f2dc99bca4e40558d05a283e99ce7dded7f1