Disable PT based demuxing if MID header extension is present.
This CL implements PT based demuxing as soon as a MID header extension
is present anywhere in the SDP, not caring if it was in the offer or
answer. In other words, either the offerer or the answerer has the
ability to turn off PT based demuxing and a spec-compliant endpoint
would default it to off.

- Want PT-based demux + early media? You need to SDP munge the offer.
- Want PT-based demux but not early media? It is sufficient for the
  answerer to remove the MID header extensions.
- Want to use MID or SSRC demuxing? Great, you don't need PT-based demux
  and won't mind that we turned it off for you.

The reason for disabling PT demux at offer time if MID is present in
the offer instead of waiting for the SDP answer to see whether or not
it was removed is because by the time the SDP answer arrives with MID
present, if PT was enabled then early media could already have caused
incorrect mappings and the spec says ssrcs are not allowed to change
m= section so this is dangerous.

This change only affects Unified Plan and bundled groups. Existing test
coverage ensuring we do not break legacy endpoints:
[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/webrtc/pc/peer_connection_integrationtest.cc;l=1156
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/external/wpt/webrtc/protocol/rtp-demuxing.html;l=59

A kill-switch is added just to be safe.

Bug: webrtc:12814
Change-Id: I807a82a543325753633aaef698e06cb4c9dfebaa