Disable PT based demuxing if MID header extension is present.
This CL disables PT based demuxing as soon as a MID header extension
is present anywhere in the SDP, not caring if it was in the offer or
answer. In other words, either the offerer or the answerer has the
ability to turn off PT based demuxing and a spec-compliant endpoint
would default it to off.

- Want PT-based demux + early media? You need to SDP munge the offer.
- Want PT-based demux but not early media? It is sufficient for the
  answerer to remove the MID header extensions.
- Want to use MID or SSRC demuxing? Great, you don't need PT-based demux
  and won't mind that we turned it off for you.

The reason for disabling PT demux at offer time (if MID is present)
instead of waiting for the SDP answer is because by the time the SDP
answer arrives, early media could have triggered PT demux and caused
incorrect mappings. The safe thing is to assume a spec-compliant
endpoint until proven otherwise.

This change only affects Unified Plan and bundled groups. Existing test
coverage ensuring we do not break legacy endpoints:
[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/webrtc/pc/peer_connection_integrationtest.cc;l=1156
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/external/wpt/webrtc/protocol/rtp-demuxing.html;l=59

UnsignaledStreamTest is also updated to test the interesting setups.
A kill-switch is added in case we want to disable this change.

Bug: webrtc:12814
Change-Id: I807a82a543325753633aaef698e06cb4c9dfebaa