Make WebRtcVideoChannel::SetSend not block on UpdateActiveSimulcastLayers
During renegotiation, UpdateActiveSimulcastLayers() gets called as part
of WebRtcVideoSendStream::UpdateSendState. This CL changes
UpdateActiveSimulcastLayers to not synchronize with the transport's
task queue to wait for the changes to get applied.

During renegotiation, especially with multiple channels, this
synchronization will be quite costly, so we should aim to get
rid of the `thread_sync_event_` in VideoSendStream.

Bug: webrtc:12840
Change-Id: Idb48d29b6b8382881c7c1e6f1d0f5e708dbca30f