Add a DLOG to RtpPacket::SetExtension.
This log is useful when searching the WebRTC log to check if
there is a problem with an RTP Header Extension (e.g. was it actually
set?).

DLOG to avoid to increase logging and binary size, it is easy to
enable in release builds by -DDLOG_ALWAYS_ON.

Example:
(rtp_packet.h:249): RTP Header Extension 16 \
    (urn:ietf:params:rtp-hdrext:sdes:mid) was not set

Bug: None
Change-Id: Ia112eec475b164230ab7114d2cc3ec1080041e57