vss: Optionally using deterministic randomness
The virtual socket server may delay packets based on parameters, but
ultimately, that delay uses randomness. It also allows setting a drop
probability, which also uses randomness to decide if a packet should be
dropped.

Before this commit, it used the libc's rand() function for that, but
that makes tests not very reproducible. This can of course be debated
how much you should rely on it, but to reduce flakiness, tests can with
this commit opt in to making VSS more deterministic, by specifying which
seed to use for the PRNG that will be used instead of lib's rand().

This is a preparation CL for adding throughput tests for dcSCTP.

Bug: webrtc:12943
Change-Id: I1750d2afbdf5d6240b799629f1ad702d9ed1a2aa