dcsctp: Allow specifying clock granularity
This is mentioned in
https://datatracker.ietf.org/doc/html/rfc4960#section-6.3.1 and further
described in https://datatracker.ietf.org/doc/html/rfc6298#section-4.

A default value of 10 ms is used, which is identical to what usrsctp has
chosen when running in user-space.

This should possibly be set much higher, to account for the delayed ack
timeout (ATO) of the peer (as that can be seen as a very high clock
granularity). That one is set to 200ms by default in many clients, so a
reasonable default limit could be set to 210 ms.

Bug: webrtc:12943
Change-Id: Ifc217daa390850520da8b3beb0ef214181ff8c4e