dcsctp: Detect the peer SCTP implementation
This is not explicitly signaled, so heuristics are used. These are not
guaranteed to come to the correct conclusion, and the data is not always
available.

It's to be used for clients to record metrics. The behavior of dcSCTP
will not change depending on the assumed implementation. Only by
explicitly signaled capabilities.

Bug: webrtc:13216
Change-Id: I2f58054d17d53d947ed5845df7a08f974d42f918