AGC2: use only one headroom parameter
Instead of using two different headroom parameters, namely one to
specify the target level and one to estimate the crest factor, use
a single parameter so that tuning becomes easier.

This CL *changes* the behavior of the adaptive digital controller and
therefore the expectation of some unit tests have been adapted.

Bug: webrtc:7494
Change-Id: I0a2a710bbede0caa53938090a004d185fdefaeb9